+++ protected/humhub/config/__autocomplete.php
@@ -39,6 +39,7 @@ class Yii
  * @property \yii\mutex\Mutex $mutex
  * @property \yii\web\User|__WebUser $user
  * @property \yii\caching\ArrayCache $runtimeCache
+ * @property \humhub\components\InstallationState $installationState
  */
 class __Application
 {