+++ protected/humhub/modules/notification/targets/MailTarget.php
@@ -8,6 +8,7 @@
 
 namespace humhub\modules\notification\targets;
 
+use humhub\components\InstallationState;
 use humhub\modules\notification\components\BaseNotification;
 use humhub\modules\user\models\User;
 use Yii;
@@ -89,6 +90,6 @@ class MailTarget extends BaseTarget
     public function isActive(User $user = null)
     {
         // Do not send mail notifications for example content during installlation.
-        return parent::isActive() && Yii::$app->isInstalled();
+        return parent::isActive() && Yii::$app->installationState->hasState(InstallationState::STATE_INSTALLED);
     }
 }